finalizando o CRUD

+++ Produtos/Models/Produto.cs
@@ -7,7 +7,7 @@ namespace Produtos.Models
         [Key]
         public int Id { get; set; }
         public string? Nome { get; set; }
-        public float Preco { get; set; }
+        public double Preco { get; set; } 
         public string? Descricao { get; set; }
         public DateTime DataRegistro { get; set; } = DateTime.Now;
     }
